Fixed textbox not updating value
